Improve performance
Eliminate the block/foldl structure and replace it with an uglier (but
faster) single function that will build the blocks and run the MD5Update
routine in a single traversal.  This means there is one less comparison
checking that we have blockSizeBytes in our bytestring.

darcs-hash:20080622000458-cef97-2151968783655f68440708133e69bf0491dcbd84.gz