Don't force inlining of long form of getNthWord
`getNthWord` has two implementations. While the implementation used when `FastWordExtract` is set is quite lean and perhaps worth inlining, the other implementation (used, among other cases, in the case of ARM targets) is far too large. This results in the simplifier exploding. Don't require inlining of the non-`FastWordExtract` case.