Export a md5DigestBytes function to get access to the raw bytes
The existing Show instance provides a convenient way to get the hex
representation and the Binary and Serialise instances also provide
access to the raw bytes. That said, the documentation does not make
clear that this is what the Binary and Serialise instances do.

So this patch adds md5DigestBytes :: MD5Digest -> B.ByteString
and haddock docs to point out the other options, of Show, Binary or
Serialise.